Fix event processing for GPIOx:y cop commands
Added code to parse both arguments to the GPIO commands.  Now both cli invocation of the command and event driven by RPT_RXKEYED method work as expected.  Typical application is for USB sound card interfaces which have a COS LED signal.  This brings functionality on par with HamVoIP for this feature and now works as documented.