Added missing build dependencies
On a minimal installation of Debian (Docker: debian:buster-slim) the following packages are required to build:

* devscripts
* doxygen
* graphviz
* gsfonts

To better track which packages are needed, I've also sorted the list alphabetically.

Note that I've only added to this list, not checked to see if all packages are required. If I spot any surplus packages I'll add a separate pull request.